Entrega proyecto final

+++ src/components/ItemDetailContainer/ItemDetailContainer.jsx
@@ -3,6 +3,7 @@ import ItemDetail from "../ItemDetail/ItemDetail";
 import { useParams } from 'react-router-dom'
 import Container from 'react-bootstrap/Container'
 import getFirestore from "../../Firebase/firebase";
+import Loader from "../Loader/Loader";
 
 const ItemDetailContainer = () => {
     const [item, setItem] = useState({});
@@ -12,22 +13,21 @@ const ItemDetailContainer = () => {
     useEffect(() => {
         const db = getFirestore();
         const products = db.collection('products');
-        const item = products.doc(id);
-        item.get().then((doc) => {
-            setItem({ id: doc.id, ...doc.data() });
+        const itemFirestore = products.doc(id);
+        itemFirestore.get().then((doc) => {
+            if (!doc.exists) {
+                console.log('No existen productos')
+                setItem({});
+            } else {
+                setItem({ id: doc.id, ...doc.data() });
+            }
         })
             .catch(err => console.log(err))
             .finally(() => setLoading(false))
     }, [id])
 
     return (
-        <div>  {loading ? (
-
-            <div className='d-flex align-items-center container' style={{ height: '50vh' }}>
-                <h2>Cargando detalle...</h2>
-                <div className="spinner-border ms-auto" role="status" aria-hidden="true"></div>
-            </div>
-
+        <div>  {loading ? ( <Loader />
         ) : (
             <Container>
                 <ItemDetail item={item} />
